requirements.txt
torch
torchvision
transformers
Pillow
opencv-python
numpy
pandas
qdrant-client
fastapi
uvicorn
python-multipart
rembg
pydantic
python-dotenv
aiofiles
tqdm
pytest
requests
PyPDF2
pdf2image
pymupdf
pytesseract
pymysql
email-validator